requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
redis>=5.0.0
orjson>=3.9.10
aiosqlite>=0.19.0
typing-extensions
aiohttp
colorama 